requirements.txt
biopython==1.78
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
mysqlclient==2.0.1
nltk==3.5
numpy==1.19.4
openpyxl==3.0.5
pandas==1.1.4
scikit-learn==0.23.2
uWSGI==2.0.19.1
